models update

+++ user/models.py
@@ -9,7 +9,7 @@ class User(AbstractUser):
         ('executor', 'Executor'),
     )
     role = models.CharField(choices=ROLE, max_length=8)
-    balance = models.PositiveIntegerField(default=0)
+    balance = models.DecimalField(max_digits=7, decimal_places=2, default=0)
     email = models.EmailField(unique=True, null=False)
 
     def __str__(self):
